requirements.txt
discord.py
imapclient
openai
python-dotenv
python-decouple
psutil
colorama